+++ archive/hardhat-ignition-examples/complete/contracts/BasicLibrary.sol
@@ -1,8 +1,0 @@
-// SPDX-License-Identifier: UNLICENSED
-pragma solidity ^0.8.9;
-
-library BasicLibrary {
-  function addTwo(uint self) public pure returns (uint) {
-    return self + 2;
-  }
-}