+++ archive/hardhat-verify/test/fixture-projects/hardhat-project/libraries.js
@@ -1,4 +1,0 @@
-module.exports = {
-  NormalLib: "0x9fE46736679d2D9a65F0992F2272dE9f3c7fa6e0",
-  ConstructorLib: "0xCf7Ed3AccA5a467e9e704C703E8D87F634fB0Fc9",
-};