+++ archive/hardhat-vyper/test/fixture-projects/old-vyper-versions/old-vyper-version.js
@@ -1,5 +1,0 @@
-require("../../../src/index");
-
-module.exports = {
-  vyper: "0.1.0-beta.15",
-};